+++ python/tvm/script/tir/__init__.pyi
@@ -28,12 +28,12 @@ from typing import (
     List,
     Mapping,
 )
+from numbers import Number
 
 from tvm.tir.function import PrimFunc
 from tvm.tir import PrimExpr, Range, IterVar, Var
 from tvm.runtime import Object
 from .node import BufferSlice
-from . import axis
 from .ty import ConcreteType
 
 """
@@ -51,19 +51,19 @@ class Ptr: ...
 Variables and constants
 """
 
-def bool(imm: int) -> PrimExpr: ...
-def int8(imm: int) -> PrimExpr: ...
-def int16(imm: int) -> PrimExpr: ...
-def int32(imm: int) -> PrimExpr: ...
-def int64(imm: int) -> PrimExpr: ...
-def uint8(imm: int) -> PrimExpr: ...
-def uint16(imm: int) -> PrimExpr: ...
-def uint32(imm: int) -> PrimExpr: ...
-def uint64(imm: int) -> PrimExpr: ...
-def float8(imm: int) -> PrimExpr: ...
-def float16(imm: int) -> PrimExpr: ...
-def float32(imm: int) -> PrimExpr: ...
-def float64(imm: int) -> PrimExpr: ...
+def bool(imm) -> PrimExpr: ...
+def int8(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def int16(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def int32(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def int64(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def uint8(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def uint16(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def uint32(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def uint64(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def float8(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def float16(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def float32(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
+def float64(imm: Union[PrimExpr, Number]) -> PrimExpr: ...
 
 """
 Intrinsic
@@ -116,15 +116,6 @@ def sqrt(x: PrimExpr) -> PrimExpr: ...
 def rsqrt(x: PrimExpr) -> PrimExpr: ...
 
 """
-Axis
-"""
-
-def reduce_axis(begin: Union[PrimExpr, int], end: Union[PrimExpr, int]) -> IterVar: ...
-def range(begin: Union[PrimExpr, int], end: Union[PrimExpr, int]) -> IterVar: ...
-def scan_axis(begin: Union[PrimExpr, int], end: Union[PrimExpr, int]) -> IterVar: ...
-def opaque_axis(begin: Union[PrimExpr, int], end: Union[PrimExpr, int]) -> IterVar: ...
-
-"""
 special_stmt - Buffers
 """
 
@@ -175,11 +166,21 @@ def block_attr(attrs: Mapping[str, Object]) -> None: ...
 special_stmt - Axis
 """
 
-def axis_spatial(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
-def axis_reduce(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
-def axis_scan(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
-def axis_opaque(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
-def axis_remap(iter_types: str, loop_vars: List[Var]) -> IterVar: ...
+class axis:
+    @staticmethod
+    def spatial(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
+    @staticmethod
+    def S(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
+    @staticmethod
+    def reduce(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
+    @staticmethod
+    def R(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
+    @staticmethod
+    def scan(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
+    @staticmethod
+    def opaque(dom: Union[PrimExpr, Tuple[PrimExpr, PrimExpr]], value: PrimExpr) -> IterVar: ...
+    @staticmethod
+    def remap(iter_types: str, loop_vars: List[Var]) -> IterVar: ...
 
 """
 special_stmt - Annotations
